-NMT_완성 -기타 코드 작성중 -CAN_ReadOD가 완성되어있지않은데, 이부분 용재한테 확인

+++ Inc/CANOpen.h
@@ -31,7 +31,7 @@ extern CO_Status CANOpen_writeOD_int16(uint8_t nodeId, uint16_t Index, uint8_t s
 extern CO_Status CANOpen_writeOD_uint8(uint8_t nodeId, uint16_t Index, uint8_t subIndex, uint8_t data, uint16_t timeout);
 extern CO_Status CANOpen_writeOD_int8(uint8_t nodeId, uint16_t Index, uint8_t subIndex, int8_t data, uint16_t timeout);
 
-extern CO_Status CANOpen_readOD(uint8_t nodeId, uint16_t Index, uint8_t subIndex, uint8_t* data, uint8_t* len, uint16_t timeout);
+extern CO_Status CANOpen_readOD(uint8_t nodeId, uint16_t Index, uint8_t subIndex, uint16_t* data, uint8_t* len, uint16_t timeout);
 
 extern void CANOpen_mappingPDO_init(CO_PDOStruct* pdo_struct);
 extern void CANOpen_mappingPDO_float(CO_PDOStruct* pdo_struct, float* data);
